Extract protect-reason text functions
