+++ src/goals/MergeDuplicates/MergeDupsStep/SenseCardContent.tsx
@@ -3,17 +3,12 @@ import { CardContent, IconButton } from "@mui/material";
 import { type ReactElement } from "react";
 import { useTranslation } from "react-i18next";
 
-import {
-  GramCatGroup,
-  type ProtectReason,
-  ReasonType,
-  type Sense,
-  Status,
-} from "api/models";
+import { GramCatGroup, type Sense, Status } from "api/models";
 import { IconButtonWithTooltip, PartOfSpeechButton } from "components/Buttons";
 import MultilineTooltipTitle from "components/MultilineTooltipTitle";
 import DomainChipsGrid from "components/WordCard/DomainChipsGrid";
 import SenseCardText from "components/WordCard/SenseCardText";
+import { protectReasonsText } from "goals/MergeDuplicates/MergeDupsStep/protectReasonUtils";
 import { combineSenses } from "goals/MergeDuplicates/Redux/reducerUtilities";
 
 interface SenseCardContentProps {
@@ -41,72 +36,12 @@ export default function SenseCardContent(
     a.id.localeCompare(b.id)
   );
 
-  const reasonText = (reason: ProtectReason): string => {
-    // Backend/Helper/LiftHelper.cs > GetProtectedReasons(LiftSense sense)
-    switch (reason.type) {
-      case ReasonType.Annotations:
-        return t("mergeDups.protectReason.annotations");
-      case ReasonType.Examples:
-        return t("mergeDups.protectReason.examples");
-      case ReasonType.Field:
-        return t("mergeDups.protectReason.field", { val: reason.value });
-      case ReasonType.GramInfoTrait:
-        return t("mergeDups.protectReason.gramInfoTrait", {
-          val: reason.value,
-        });
-      case ReasonType.Illustrations:
-        return t("mergeDups.protectReason.illustrations");
-      case ReasonType.Notes:
-        return t("mergeDups.protectReason.notesSense");
-      case ReasonType.Relations:
-        return t("mergeDups.protectReason.relations");
-      case ReasonType.Reversals:
-        return t("mergeDups.protectReason.reversal", { val: reason.value });
-      case ReasonType.Subsenses:
-        return t("mergeDups.protectReason.subsenses");
-      case ReasonType.Trait:
-        return reason.value ?? "(unknown trait)";
-      case ReasonType.TraitAnthroCode:
-        return t("mergeDups.protectReason.traitAnthroCode", {
-          val: reason.value,
-        });
-      case ReasonType.TraitDomainType:
-        return t("mergeDups.protectReason.traitDomainType", {
-          val: reason.value,
-        });
-      case ReasonType.TraitDoNotPublishIn:
-        return t("mergeDups.protectReason.traitDoNotPublishIn", {
-          val: reason.value,
-        });
-      case ReasonType.TraitPublishIn:
-        return t("mergeDups.protectReason.traitPublishIn", {
-          val: reason.value,
-        });
-      case ReasonType.TraitSenseType:
-        return t("mergeDups.protectReason.traitSenseType", {
-          val: reason.value,
-        });
-      case ReasonType.TraitStatus:
-        return t("mergeDups.protectReason.traitStatus", { val: reason.value });
-      case ReasonType.TraitUsageType:
-        return t("mergeDups.protectReason.traitUsageType", {
-          val: reason.value,
-        });
-      default:
-        throw new Error();
-    }
-  };
-
   const protectedWarning =
     !props.sidebar && sense.accessibility === Status.Protected;
   const tooltipTexts = [t("mergeDups.helpText.protectedSense")];
   const reasons = sense.protectReasons;
   if (reasons?.length) {
-    tooltipTexts.push(
-      t("mergeDups.helpText.protectedData", {
-        val: reasons.map(reasonText).join("; "),
-      })
-    );
+    tooltipTexts.push(protectReasonsText(t, [], reasons));
   }
   tooltipTexts.push(t("mergeDups.helpText.protectedSenseInfo"));
